+++ src/specs/mod.rs
@@ -20,6 +20,7 @@ mod given;
 
 mod for_artifact_detection;
 mod for_branch_prunability;
+mod for_configuration_editing;
 mod for_docker_reclaim;
 mod for_personal_triage;
 mod for_size_reporting;